fix: Move Extensions into Common where appropriate.
