+++ Assets/Plugins/Source/Editor/Utility/PackageFileUtility.cs
@@ -20,16 +20,17 @@
 * SOFTWARE.
 */
 
-using System.Collections.Generic;
-using UnityEngine;
-using System.IO;
-using System;
-using System.Text.RegularExpressions;
+
 
 namespace PlayEveryWare.EpicOnlineServices.Utility
 {
+    using System.Collections.Generic;
+    using UnityEngine;
+    using System.IO;
+    using System;
+    using System.Text.RegularExpressions;
+    using Common.Extensions;
     using Editor.Build;
-    using Extensions;
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
