+++ com.playeveryware.eos/Runtime/Core/Common/Extensions/FileInfoExtensions.cs.meta
