+++ com.playeveryware.eos/Runtime/Core/Common/Extensions/ListExtensions.cs
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-namespace PlayEveryWare.EpicOnlineServices.Extensions
+namespace PlayEveryWare.Common.Extensions
 {
     using System;
     using System.Collections.Generic;
