+++ com.playeveryware.eos/Runtime/Core/Config/ProductionEnvironments.cs
@@ -23,6 +23,7 @@
 namespace PlayEveryWare.EpicOnlineServices
 {
     using Common;
+    using PlayEveryWare.Common;
 
     /// <summary>
     /// This class contains information about the set of deployments and
